더하기,빼기,곱하기,나누기 class생성 및 main & calculator 수정

+++ app/src/main/java/com/example/calculator/Main.kt
@@ -3,11 +3,11 @@ package com.example.calculator
  fun main() {
 
      println("숫자를 입력해 주세요.")
-     var num1 = readLine()!!.toInt()
+     var num1 = readLine()!!.toDouble()
      println("연산을 입력해 주세요.(+,-,*,/,%)")
      var op = readLine()!!.toString()
      println("숫자를 입력해 주세요.")
-     var num2 = readLine()!!.toInt()
+     var num2 = readLine()!!.toDouble()
 
      var calc = Calculator()
      println("결과는 ${calc.operate(num1, num2, op)}입니다.")
